spring8-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.10.1
pydantic==2.10.3
pydantic-settings==2.6.1
google-generativeai
python-multipart==0.0.12
python-dotenv==1.0.1